Update mqtt message format (#47)
* Remove mqtt_event code


* Add metadata block to mqtt messages


* Fix UCI config read

This commit fixes how values are read from powquty config.
The uci_foreach_element 'loop' left after a continue to look
for a new section, not a new option. This caused some errors if
a option was not set, as all following(in the C code) options
where not read. The same behavior was shown, when an option
value was to long. This is no longer the case and not every
option needs to be set anymore.


* Compose static mqtt data in advance


* Fix some warnings and make meta-data completely optional


* Add package count to mqtt message


* Add optional t5060 and t1012 data


* Update Readme and provide config file options


* Use actual t1020 voltage and frequency


* Fix indention


* Fix typo


* Make mqtt_port macro all uppercase


* Rename function to config_option_is_valid


* Adjust max length for static and meta data


* Update pkg_count to new constraints


* Remove comment of old message structure

+++ powqutyd/files/src/raw_dump.c
@@ -21,11 +21,11 @@ static pthread_cond_t dump_cond;
 static pthread_mutex_t dump_mtx;
 static pthread_t raw_dump_thread;
 
-unsigned char * dump_buffer;	// [MAX_FRAME_SIZE*DUMP_BUFFER_SIZE];
-char *dump_mode; 				// [DUMP_BUFFER_SIZE];
+unsigned char * dump_buffer;		// [MAX_FRAME_SIZE*DUMP_BUFFER_SIZE];
+char *dump_mode;			// [DUMP_BUFFER_SIZE];
 long long *dump_time;			// [DUMP_BUFFER_SIZE];
-int *dump_size;					// [DUMP_BUFFER_SIZE];
-char *dump_string;				// [MAX_DUMP_STRING];
+int *dump_size;				// [DUMP_BUFFER_SIZE];
+char *dump_string;			// [MAX_DUMP_STRING];
 static volatile short new_pkt_idx=0, pkt_to_dump_idx=0;
 static volatile unsigned int stop_raw_dump_run = 0;
 unsigned short last_frame_idx =0;
@@ -75,7 +75,8 @@ void dump_raw_packet(unsigned char* frame, int read_size, char mode) {
 }
 
 void* raw_dump_run(void* args) {
-	printf("DEBUG:\tDump Thread has started with Thread Id: %ld\n", pthread_self());
+	printf("DEBUG:\tDump Thread has started with Thread Id: %lu\n",
+		(long unsigned int)pthread_self());
 	while(!stop_raw_dump_run) {
 		pthread_mutex_lock(&dump_mtx);
 		pthread_cond_wait(&dump_cond,&dump_mtx);
